update: prepare.handler.js 내부 로직 함수화 및 분리
